% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. Numeros Primos.R
\docType{data}
\name{Impresion}
\alias{Impresion}
\title{Prime Number Program}
\format{
An object of class \code{logical} of length 1.
}
\usage{
Impresion
}
\arguments{
\item{Modulo}{}
}
\value{
returns integer values

[1] 1
[1] 2
[1] 3
[1] ...
[1] 89
[1] 97
}
\description{
Prints the prime numbers from 1 to 100, each on a separator line
}
\keyword{datasets}
