% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. Retrieve_answer.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{Programa para impresion de resultados}
\usage{
retrieve_answer(x)
}
\arguments{
\item{x}{debe ser un parametro numerico}
}
\value{
Retorna uno de los resultados de cada punto dela guia
}
\description{
Imprime los resultados de la guia de exploración de R
para ciencia de datos dependiendo del numero que coloques
}
\examples{
Retrieve_answer(1)
Retrieve_answer(6)
}
